Refactoring runtimeName test
